Supporing find format for file walking
